Extract contracts to Cog\Contracts\YouTrack package
