+++ composer.json
@@ -48,6 +48,7 @@
     },
     "autoload": {
         "psr-4": {
+            "Cog\\Contracts\\YouTrack\\Rest\\": "contracts/",
             "Cog\\YouTrack\\Rest\\": "src/"
         }
     },
