+++ src/Authorizer/CookieAuthorizer.php
@@ -13,9 +13,9 @@ declare(strict_types=1);
 
 namespace Cog\YouTrack\Rest\Authorizer;
 
-use Cog\YouTrack\Rest\Authenticator\Contracts\Authenticator as AuthenticatorContract;
-use Cog\YouTrack\Rest\Authorizer\Contracts\Authorizer as AuthorizerContract;
-use Cog\YouTrack\Rest\Client\Contracts\Client as ClientContract;
+use Cog\Contracts\YouTrack\Rest\Authenticator\Authenticator as AuthenticatorContract;
+use Cog\Contracts\YouTrack\Rest\Authorizer\Authorizer as AuthorizerContract;
+use Cog\Contracts\YouTrack\Rest\Client\Client as ClientContract;
 
 /**
  * Class CookieAuthorizer.
@@ -25,14 +25,14 @@ use Cog\YouTrack\Rest\Client\Contracts\Client as ClientContract;
 class CookieAuthorizer implements AuthorizerContract
 {
     /**
-     * @var \Cog\YouTrack\Rest\Authenticator\Contracts\Authenticator
+     * @var \Cog\Contracts\YouTrack\Rest\Authenticator\Authenticator
      */
     private $authenticator;
 
     /**
      * CookieAuthorizer constructor.
      *
-     * @param \Cog\YouTrack\Rest\Authenticator\Contracts\Authenticator $authenticator
+     * @param \Cog\Contracts\YouTrack\Rest\Authenticator\Authenticator $authenticator
      */
     public function __construct(AuthenticatorContract $authenticator)
     {
@@ -42,7 +42,7 @@ class CookieAuthorizer implements AuthorizerContract
     /**
      * Returns authorization headers.
      *
-     * @param \Cog\YouTrack\Rest\Client\Contracts\Client $client
+     * @param \Cog\Contracts\YouTrack\Rest\Client\Client $client
      * @return void
      */
     public function appendHeadersTo(ClientContract $client): void
